feature-BJS2-25762
Create ObjectMapperConfig
