+++ src/main/java/faang/school/notificationservice/NotificationServiceApp.java
@@ -17,11 +17,4 @@ public class NotificationServiceApp {
                 .bannerMode(Banner.Mode.OFF)
                 .run(args);
     }
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.registerModule(new JavaTimeModule());
-        return objectMapper;
-    }
 }
